use only a single compute_md5sum function
